Trigger horovod allreduce in eager mode
Previously, we relied on horovod's DistributedOptimizer to do
average the gradients. However, the optimizer.get_gradients code
path was not even triggered in eager mode or when a distribution
strategy is used. The new approach taken here is to always run
in eager mode and not use a distribution strategy, and do the
allreduce manually ourselves instead of relying on an optimizer.