Explicitly deallocate tensors between virtual nodes
This commit aims to fix the problem that memory was not released
until the end of the execution function. This causes OOMs when
we run multiple virtual nodes on the same device, because the
intermediate tensors (mostly activations) were never deallocated.
To address this issue, this commit introduces a new op kernel
that explicitly deallocates these tensors after each virtual
node is run.

Note: this implementation doesn't work with the ResNet-50 on
ImageNet workload yet. The next immediate goal is to debug this.