Add control deps instead of explicitly deallocating memory
The root cause of OOMs was not because memory wasn't being freed.
It was because the previous code let multiple virtual nodes run
at the same time. This is caused by the lack of control
dependencies across loop iterations, i.e. the looping mechanism
did not exist in the underlying graph.

This commit fixes the root cause of the problem by introducing
such control dependencies between iterations of the loop. With
these changes, we can now run BERT fine-tuning on the MRPC task
with a batch size of 256 on a single GPU with 8 virtual nodes
assigned to the GPU. The memory usage of doing so is similar to
running with a batch size of 32 on the same GPU.

A future commit can revert the changes that allowed us to
explicitly deallocate a tensor's memory.