Refactor keras API to accept multiple datasets
This commit allows the user to pass in multiple datasets into
`model.fit`, one for each virtual node, instead of just one.
When processing a virtual node, tensorflow will now draw from
the specific dataset iterator that corresponds to that virtual
node, enabling us to "transfer" virtual nodes in the future
when elasticity is implemented.