Optionally synchronize gradients with Horovod
This commit adds support for using Horovod to synchronize
gradients across workers. Note that within a worker the native
distribution strategy is still used to first synchronize
gradients across the devices assigned to the worker, before
passing the gradients to Horovod.

There are two potential issues with the current implementation:

1. Performance: Horovod allreduce currently runs outside the graph.
This means we need to run the existing graph first, pass the
gradients to Horovod, and run another op to apply the gradients.
We chose this so allreduce is a narrow-waist function can be
replaced quickly when we introduce elasticity later, without
rebuilding the entire graph. However, we pay an overhead for
this design.

2. End of sequence exception on model.evaluate(). This happens only
when Horovod is used for some reason. Must be a bug somewhere.