Allow num_virtual_nodes to change across batches
This commit lays the foundation for elasticity. Previously,
changing the number of virtual nodes would require rebuilding
the entire graph. This is because the graph was created by
stitching together operations N times instead of using the
while loop operator in TensorFlow. This commit refactors the
train and test functions to use the `tf.while_loop` operator.

An additional benefit of this approach is it cuts the graph
building time significantly, since now the size of the graph
does not scale with the number of virtual nodes. This is
significant for a large number of virtual nodes, e.g. 64.

An early version of this commit attempted to refactor this
code using the new looping mechanism in autograph. However,
all attempts were met with bizarre error messages that did
not convey the root cause, so we resorted to using the legacy
`tf.while_loop` instead.