Fix GPU OOM in distributed mode
The culprit was `tf.cast`. Previously we tried to cast num
virtual nodes (a tf.constant) into the same data type as the
gradients. Now we directly pass num virtual nodes as an int
to avoid the cast. Note that this means every time we change
the number of virtual nodes we have to retrace the function.