Sum, not average, local losses (gradients) for HT
HT = heterogeneous training

We do this because averaging local losses would lead to incorrect
gradients. E.g. suppose we have 2 GPUs and a global batch size of
64. GPU X is assigned 48 examples while GPU Y is assigned 16
examples. If we average local losses, which is the default, we
will end up with:

((X / 48) + (Y / 16)) / 2 = (X + 3Y) / 96

But what we want is:

(X + Y) / 64

which would match the semantics of running the same batch in a
non-heterogeneous setting. Therefore, we need to sum the losses
(X + Y) and divide by the global batch size (64) instead.