Add support for heterogeneous training (round 1)
This commit adds support for uneven sharding and data processing.
This is achieved independently on each worker by skipping over
examples that are not assigned to it. The ratio of how much data
to assign to the workers in each batch is specified through a
variable we call "heterogeneous split". For e.g., 12,3,2 means
the global batch size is 17 (12+3+2) and there are three workers.
The first worker is assigned 12 examples in each batch, the
second worker is assigned 3, and the third worker 2. This means
the first worker will process examples [0-12) in the first batch,
[17-29) in the second batch and so on.

The changes here assume the dataset will be batched (after
sharding) by the per-worker batch size instead of the global
batch size. The per-worker batch is then split among the replicas
on the worker by the multi-device iterator as before. More
specifically, the rebatching that happens in DistributedIterator
is now for splitting the per-worker batch into one or more
per-replica batches.