Fix heterogeneous + virtual nodes convergence
This is an attempt to fix the convergence behavior of combining
heterogeneous training with virtual nodes. Previously, we would
average the gradients across all virtual nodes on each device,
then sum these gradients across all devices and divide the
gradients by the global batch size. This is wrong, however,
cause we are essentially double scaling the gradients!

For e.g. global batch size = 8192
GPU 0 batch size = 6144, num virtual nodes = 32
GPU 1 batch size = 2048, num virtual nodes = 32

Before, we would get ((g0 / 32) + (g1 / 32)) / 8192, where g0 is
the sum of the gradients across all virtual nodes on GPU 0 and
so on. However, the correct gradients should be (g0 + g1) / 8192,
because dividing by 8192 in the end already takes care of the
averaging. To fix this problem, this commit no longer divides
the aggregated gradients by the number of virtual nodes in the
heterogeneous case.