Fix heterogeneous + virtual nodes convergence (round 2)
This reverts b4ec231 and 1b7204c. The previous approach of
summing all losses and then dividing by the global batch size
proved to be problematic with virtual nodes. This is because
when we sum across the virtual nodes on a device, we end up
summing the regularization losses as well, but these losses
were not scaled by the global batch size. However, we can't
simply average across the virtual nodes either, because that
would by double dividing.

Instead, we try a new approach. We perform a weighted average
for gradient synchronization instead. This allows us to keep
averaging across the virtual nodes on the same device. The
scaling factor is applied when the losses are computed as in
the previous code.